requirements.txt
# Amsterdam Strain Finder — Python dependencies
httpx>=0.27
beautifulsoup4>=4.12
sqlalchemy>=2.0
anthropic>=0.40
pydantic>=2.0
fastapi>=0.115
uvicorn[standard]>=0.30
apscheduler>=3.10
Pillow>=10.0
pdf2image>=1.17
python-dotenv>=1.0
